+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.1.4
+## 0.1.x
 ### Fixed
 - `DirectoryObserver` now clears the old adapter data before creating more.
 - `DirectoryObserver` can now be filtered
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed crash when accessing data directory
 - Can now specify an external path in `DirectoryObserver`
 - Directory observer now properly observes only transactions in the source directory
+- Directory observer can now sort files with a custom comparator
+- Directory observer now records sync state with remote device
 
 ## 0.1.0
 ### Added